test: group commands tests similar to `system/Commands/`
