+++ tests/system/Commands/DatabaseCommandsTest.php
@@ -18,6 +18,8 @@ use CodeIgniter\Test\StreamFilterTrait;
 use PHPUnit\Framework\Attributes\Group;
 
 /**
+ * @todo To figure out how to transfer this test to `tests/system/Commands/Database/` without breaking DatabaseLive group.
+ *
  * @internal
  */
 #[Group('DatabaseLive')]
