+++ tests/system/Commands/Generators/ConfigGeneratorTest.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * the LICENSE file that was distributed with this source code.
  */
 
-namespace CodeIgniter\Commands;
+namespace CodeIgniter\Commands\Generators;
 
 use CodeIgniter\Test\CIUnitTestCase;
 use CodeIgniter\Test\StreamFilterTrait;
@@ -21,34 +21,28 @@ use PHPUnit\Framework\Attributes\Group;
  * @internal
  */
 #[Group('Others')]
-final class ValidationGeneratorTest extends CIUnitTestCase
+final class ConfigGeneratorTest extends CIUnitTestCase
 {
     use StreamFilterTrait;
 
     protected function tearDown(): void
     {
-        parent::tearDown();
-
         $result = str_replace(["\033[0;32m", "\033[0m", "\n"], '', $this->getStreamFilterBuffer());
         $file   = str_replace('APPPATH' . DIRECTORY_SEPARATOR, APPPATH, trim(substr($result, 14)));
-        $dir    = dirname($file);
         if (is_file($file)) {
             unlink($file);
         }
-        if (is_dir($dir)) {
-            rmdir($dir);
-        }
     }
 
-    public function testGenerateValidation(): void
+    public function testGenerateConfig(): void
     {
-        command('make:validation user');
-        $this->assertFileExists(APPPATH . 'Validation/User.php');
+        command('make:config auth');
+        $this->assertFileExists(APPPATH . 'Config/Auth.php');
     }
 
-    public function testGenerateValidationWithOptionSuffix(): void
+    public function testGenerateConfigWithOptionSuffix(): void
     {
-        command('make:validation admin -suffix');
-        $this->assertFileExists(APPPATH . 'Validation/AdminValidation.php');
+        command('make:config auth -suffix');
+        $this->assertFileExists(APPPATH . 'Config/AuthConfig.php');
     }
 }
