+++ tests/system/Commands/Utilities/EnvironmentCommandTest.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * the LICENSE file that was distributed with this source code.
  */
 
-namespace CodeIgniter\Commands;
+namespace CodeIgniter\Commands\Utilities;
 
 use CodeIgniter\Config\Services;
 use CodeIgniter\Superglobals;
